requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0

# Data
tqdm>=4.65.0

# OCR (optional)
easyocr>=1.7.0

# Development
pytest>=7.0.0
